refactor: extract app home to a component
